app/requirements.txt
grpcio-tools==1.46.0
grpcio==1.46.0
pymongo==4.1.1